Update page templates and section
Update page templates and sections to work with reworked Netlify CMS config. Titles and Expandable sections are now blocks that are added in manually for the case where multiple h1 titles are needed on a page.
Styles, Do not display counter if only expandable content is only of type on page
Include id for headings and sections so can be used as anchor links